LayoutTests/platform/chromium-mac/fast/notifications/notifications-document-close-crash-expected.txt

CONSOLE MESSAGE: line 36: Uncaught Error: INVALID_STATE_ERR: DOM Exception 11
PASS
